Apply TDlib changes

+++ Unigram/Unigram/ViewModels/Settings/SettingsProxiesViewModel.cs
@@ -87,7 +87,7 @@ namespace Unigram.ViewModels.Settings
             }
         }
 
-        private void Handle(ConnectionState state, int enabledProxyId)
+        private void Handle(ConnectionState state, long enabledProxyId)
         {
             foreach (var item in Items)
             {
